rag-backend/requirements.txt
flask
flask-cors
elasticsearch
openai
python-dotenv